+++ vue/components/organisms/index.js
@@ -1,4 +1,4 @@
-import { RipeConfigurator } from "./configurator/configurator.vue";
+import { RipeConfigurator } from "./ripe-configurator/ripe-configurator.vue";
 
 const install = Vue => {
     Vue.component("ripe-configurator", RipeConfigurator);
