+++ vue/components/organisms/ripe-configurator/ripe-configurator.stories.js
