+++ vue/components/organisms/ripe-configurator/ripe-configurator.vue
