+++ vue/test/components/configurator/ripe-configurator.test.js
@@ -1,7 +1,7 @@
 const assert = require("assert");
 const base = require("../../base");
 
-describe("Configurator", () => {
+describe("RipeConfigurator", () => {
     it("should instantiate the component", () => {
         const component = base.getComponent("RipeConfigurator", {
             props: { brand: "dummy", model: "cube", version: 52, size: 1000 }
